requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-adk
litellm>=1.75.5